requirements.txt
annotated-types==0.7.0
anyio==4.12.0
async-timeout==4.0.3
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.81
langchain-text-splitters==0.3.11
langsmith==0.4.37
mido==1.3.3
numpy==2.0.2
openai==2.14.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-osc==1.8.3
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
zstandard==0.25.0
